Visual studio Code (HTML,CSS and Javascript)
In this project, I create a weather app for a website using HTML, CSS, and Javascript. Here we will display the weather information on our website from a free API using Javascript. In this weather app, we will add a search box where the user can enter the city name and get the weather information for that particular key. We will get the current weather data from the FESS API key with the help of Open Weather Map and display the weather information like temperature, weather condition, humidity, and wind speed according to the city on our website or app.